add JdbcDogDao.

+++ src/main/java/model/Dog.java
@@ -4,7 +4,10 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.Accessors;
 
-import javax.validation.constraints.*;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Past;
+import javax.validation.constraints.Positive;
+import javax.validation.constraints.Size;
 import java.util.Date;
 
 @Getter
@@ -24,5 +27,5 @@ public class Dog {
     Integer height;
 
     @Positive
-    Double weight;
+    Integer weight;
 }
